pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "precise-mrd"
version = "0.1.0"
description = "ctDNA/UMI MRD simulator + caller with deterministic error modeling and LoB/LoD estimation"
authors = [
    {name = "Precise MRD Team", email = "team@precise-mrd.org"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["ctdna", "mrd", "minimal-residual-disease", "umi", "bioinformatics", "cancer-genomics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "statsmodels>=0.14.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "typer>=0.9.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/precise-mrd/precise-mrd-mini"
Documentation = "https://precise-mrd.github.io/precise-mrd-mini/"
Repository = "https://github.com/precise-mrd/precise-mrd-mini"
Issues = "https://github.com/precise-mrd/precise-mrd-mini/issues"
Changelog = "https://github.com/precise-mrd/precise-mrd-mini/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "hypothesis>=6.80.0",
    "ruff>=0.0.280",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "coverage-badge>=1.1.0",
    "bandit>=1.7.5",
    "pip-tools>=7.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.2.0",
    "mkdocstrings[python]>=0.23.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.6",
]
rust = [
    "maturin>=1.0.0",
]

[project.scripts]
precise-mrd = "precise_mrd.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --strict-config"

[tool.coverage.run]
source = ["src/precise_mrd"]
omit = ["*/tests/*", "*/rust_ext/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]